refactoring docker-requirement out of the way
