+++ tests/test_bitcoind.py
@@ -2,7 +2,6 @@ import logging
 import os
 from cryptoadvance.specter.util.shell import which
 from cryptoadvance.specter.bitcoind import BitcoindPlainController
-from cryptoadvance.specter.bitcoind import BitcoindDockerController
 from cryptoadvance.specter.bitcoind import fetch_wallet_addresses_for_mining
 
 
@@ -13,6 +12,8 @@ def test_bitcoinddocker_running(caplog, docker, request):
     caplog.set_level(logging.DEBUG, logger="cryptoadvance.specter")
     requested_version = request.config.getoption("--bitcoind-version")
     if docker:
+        from cryptoadvance.specter.bitcoind_docker import BitcoindDockerController
+
         my_bitcoind = BitcoindDockerController(
             rpcport=18999, docker_tag=requested_version
         )  # completly different port to not interfere
